통합테스트를 컨트롤러 유닛테스트로 변환
