+++ AccountApi/src/main/java/com/nhnacademy/edu/minidooray/accountapi/domain/User.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.edu.minidooray.AccountApi.domain;
+package com.nhnacademy.edu.minidooray.accountapi.domain;
 
 import java.time.LocalDate;
 import javax.persistence.Column;
