+++ AccountApi/src/main/java/com/nhnacademy/edu/minidooray/accountapi/model/response/hasAccountResponse.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.edu.minidooray.AccountApi.model.response;
+package com.nhnacademy.edu.minidooray.accountapi.model.response;
 
 import lombok.Value;
