+++ AccountApi/src/test/java/com/nhnacademy/edu/minidooray/accountapi/AccountApiApplicationTests.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.edu.minidooray.AccountApi;
+package com.nhnacademy.edu.minidooray.accountapi;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -6,8 +6,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest
 class AccountApiApplicationTests {
 
-	@Test
-	void contextLoads() {
-	}
+    @Test
+    void contextLoads() {
+    }
 
 }
